[logisim-banked-memory] Add new attribute in AutoRAM & AutoROM.
New attribute shows if you can autoload image file after reloading the circuit.
