+++ logisim/logisim-banked-memory/src/main/java/org/cdm/logisim/memory/BankedRomAttributes.java
@@ -9,13 +9,14 @@ import java.util.WeakHashMap;
 
 import com.cburch.logisim.data.AbstractAttributeSet;
 import com.cburch.logisim.data.Attribute;
+import com.cburch.logisim.data.AttributeOption;
 import com.cburch.logisim.data.BitWidth;
 import com.cburch.logisim.gui.hex.HexFrame;
 import com.cburch.logisim.proj.Project;
 
 class BankedRomAttributes extends AbstractAttributeSet {
     private static List<Attribute<?>> ATTRIBUTES = Arrays.asList(new Attribute<?>[]{
-            BankedMem.ADDR_ATTR, BankedROM.CONTENTS_ATTR, BankedMem.PATH_ATTRIBUTE
+            BankedMem.ADDR_ATTR, BankedROM.CONTENTS_ATTR, BankedMem.PATH_ATTRIBUTE, BankedMem.LOAD_FROM_IMAGE_FILE
     });
 
     private static WeakHashMap<BankedMemContents, BankedRomContentsListener> listenerRegistry
@@ -45,6 +46,7 @@ class BankedRomAttributes extends AbstractAttributeSet {
     private BitWidth dataBits = BitWidth.create(8);
     private BankedMemContents contents;
     private String autoRom = "";
+    private AttributeOption loadFromImageFile = BankedMem.NO_IMAGE_FILE;
 
     BankedRomAttributes() {
         contents = BankedMemContents.create(addrBits.getWidth(), dataBits.getWidth());
@@ -61,6 +63,7 @@ class BankedRomAttributes extends AbstractAttributeSet {
         d.dataBits = dataBits;
         d.contents = contents.clone();
         d.autoRom = autoRom;
+        d.loadFromImageFile = loadFromImageFile;
     }
 
     @Override
@@ -75,6 +78,7 @@ class BankedRomAttributes extends AbstractAttributeSet {
         if (attr == BankedMem.DATA_ATTR) return (V) dataBits;
         if (attr == BankedROM.CONTENTS_ATTR) return (V) contents;
         if (attr == BankedMem.PATH_ATTRIBUTE) return (V) autoRom;
+        if (attr == BankedMem.LOAD_FROM_IMAGE_FILE) return (V) loadFromImageFile;
         return null;
     }
 
@@ -90,6 +94,8 @@ class BankedRomAttributes extends AbstractAttributeSet {
             contents = (BankedMemContents) value;
         } else if (attr == BankedMem.PATH_ATTRIBUTE) {
             autoRom = (String) value;
+        } else if (attr == BankedMem.LOAD_FROM_IMAGE_FILE){
+            loadFromImageFile = (AttributeOption) value;
         }
         fireAttributeValueChanged(attr, value);
     }
